Split and delete store list command; update store search command
